move calculator to own component
